Add authenticated route middleware
