+++ client/src/routes/_authenticated/clothing-management/items/$clothingItemId/edit.tsx
@@ -1,0 +1,18 @@
+import { createFileRoute } from "@tanstack/react-router"
+
+import EditClothingItemPage from "#/clothing/components/items/edit/EditClothingItemPage"
+import RoleGuard from "#/components/base/RoleGuard"
+
+export const Route = createFileRoute(
+  "/_authenticated/clothing-management/items/$clothingItemId/edit",
+)({
+  component: EditClothingItemRoute,
+})
+
+function EditClothingItemRoute() {
+  return (
+    <RoleGuard allowedRoles={["KLEIDERWART"]}>
+      <EditClothingItemPage />
+    </RoleGuard>
+  )
+}
