+++ client/src/routes/_authenticated/clothing-management/items/batch.tsx
@@ -1,0 +1,11 @@
+import { createFileRoute } from "@tanstack/react-router"
+
+import ClothingItemBatchImportPage from "#/clothing/components/items/batch-import/ClothingItemBatchImportPage"
+
+export const Route = createFileRoute("/_authenticated/clothing-management/items/batch")({
+  component: ClothingItemBatchImportRoute,
+})
+
+function ClothingItemBatchImportRoute() {
+  return <ClothingItemBatchImportPage />
+}
