+++ client/src/routes/_authenticated/clothing-management/items/index.tsx
@@ -1,0 +1,11 @@
+import { createFileRoute } from "@tanstack/react-router"
+
+import ClothingItemsPage from "#/clothing/components/items/list/ClothingItemsPage"
+
+export const Route = createFileRoute("/_authenticated/clothing-management/items/")({
+  component: ClothingItemsRoute,
+})
+
+function ClothingItemsRoute() {
+  return <ClothingItemsPage />
+}
