+++ client/src/routes/_authenticated/clothing-management/items/new.tsx
@@ -1,0 +1,16 @@
+import { createFileRoute } from "@tanstack/react-router"
+
+import RoleGuard from "#/components/base/RoleGuard"
+import ClothingItemForm from "#/clothing/components/shared/ClothingItemForm.tsx"
+
+export const Route = createFileRoute("/_authenticated/clothing-management/items/new")({
+  component: CreateClothingItemRoute,
+})
+
+function CreateClothingItemRoute() {
+  return (
+    <RoleGuard allowedRoles={["KLEIDERWART"]}>
+      <ClothingItemForm existingItem={undefined} />
+    </RoleGuard>
+  )
+}
