+++ client/src/routes/_authenticated/clothing-management/locations/$clothingLocationId/edit.tsx
@@ -1,0 +1,13 @@
+import { createFileRoute } from "@tanstack/react-router"
+
+import EditClothingLocationPage from "#/clothing/components/locations/edit/EditClothingLocationPage"
+
+export const Route = createFileRoute(
+  "/_authenticated/clothing-management/locations/$clothingLocationId/edit",
+)({
+  component: EditClothingLocationRoute,
+})
+
+function EditClothingLocationRoute() {
+  return <EditClothingLocationPage />
+}
