+++ client/src/routes/_authenticated/clothing-management/locations/new.tsx
@@ -1,0 +1,11 @@
+import { createFileRoute } from "@tanstack/react-router"
+
+import CreateClothingLocationPage from "#/clothing/components/locations/create/CreateClothingLocationPage"
+
+export const Route = createFileRoute("/_authenticated/clothing-management/locations/new")({
+  component: CreateClothingLocationRoute,
+})
+
+function CreateClothingLocationRoute() {
+  return <CreateClothingLocationPage />
+}
