+++ client/src/routes/_authenticated/clothing-management/types/$clothingTypeId/edit.tsx
@@ -1,0 +1,13 @@
+import { createFileRoute } from "@tanstack/react-router"
+
+import EditClothingTypePage from "#/clothing/components/types/edit/EditClothingTypePage"
+
+export const Route = createFileRoute(
+  "/_authenticated/clothing-management/types/$clothingTypeId/edit",
+)({
+  component: EditClothingTypeRoute,
+})
+
+function EditClothingTypeRoute() {
+  return <EditClothingTypePage />
+}
