+++ client/src/routes/_authenticated/clothing-management/types/index.tsx
@@ -1,0 +1,11 @@
+import { createFileRoute } from "@tanstack/react-router"
+
+import ClothingTypesPage from "#/clothing/components/types/list/ClothingTypesPage"
+
+export const Route = createFileRoute("/_authenticated/clothing-management/types/")({
+  component: ClothingTypesRoute,
+})
+
+function ClothingTypesRoute() {
+  return <ClothingTypesPage />
+}
