+++ client/src/routes/_authenticated/clothing-management/types/new.tsx
@@ -1,0 +1,11 @@
+import { createFileRoute } from "@tanstack/react-router"
+
+import CreateClothingTypePage from "#/clothing/components/types/create/CreateClothingTypePage"
+
+export const Route = createFileRoute("/_authenticated/clothing-management/types/new")({
+  component: CreateClothingTypeRoute,
+})
+
+function CreateClothingTypeRoute() {
+  return <CreateClothingTypePage />
+}
