+++ client/src/routes/_authenticated/pool-clothing/index.tsx
@@ -1,0 +1,11 @@
+import { createFileRoute } from "@tanstack/react-router"
+
+import PoolKlamottenPage from "#/clothing/components/pool-klamotten/PoolKlamottenPage"
+
+export const Route = createFileRoute("/_authenticated/pool-clothing/")({
+  component: PoolClothingRoute,
+})
+
+function PoolClothingRoute() {
+  return <PoolKlamottenPage />
+}
