+++ client/src/routes/_authenticated/user-management/index.tsx
@@ -1,0 +1,11 @@
+import { createFileRoute } from "@tanstack/react-router"
+
+import UsersManagementPage from "#/users/components/UsersManagementPage"
+
+export const Route = createFileRoute("/_authenticated/user-management/")({
+  component: UserManagementRoute,
+})
+
+function UserManagementRoute() {
+  return <UsersManagementPage />
+}
