--- client/src/routes/clothing-management/index.tsx
@@ -1,11 +1,0 @@
-import { createFileRoute } from "@tanstack/react-router"
-
-import ClothingManagementOverviewPage from "#/clothing/components/overview/ClothingManagementOverviewPage"
-
-export const Route = createFileRoute("/clothing-management/")({
-  component: ClothingManagementIndexRoute,
-})
-
-function ClothingManagementIndexRoute() {
-  return <ClothingManagementOverviewPage />
-}
