--- client/src/routes/clothing-management/items/$clothingItemId/edit.tsx
@@ -1,18 +1,0 @@
-import { createFileRoute } from "@tanstack/react-router"
-
-import EditClothingItemPage from "#/clothing/components/items/edit/EditClothingItemPage"
-import RoleGuard from "#/components/base/RoleGuard"
-
-export const Route = createFileRoute(
-  "/clothing-management/items/$clothingItemId/edit",
-)({
-  component: EditClothingItemRoute,
-})
-
-function EditClothingItemRoute() {
-  return (
-    <RoleGuard allowedRoles={["KLEIDERWART"]}>
-      <EditClothingItemPage />
-    </RoleGuard>
-  )
-}
