--- client/src/routes/clothing-management/locations/new.tsx
@@ -1,11 +1,0 @@
-import { createFileRoute } from "@tanstack/react-router"
-
-import CreateClothingLocationPage from "#/clothing/components/locations/create/CreateClothingLocationPage"
-
-export const Route = createFileRoute("/clothing-management/locations/new")({
-  component: CreateClothingLocationRoute,
-})
-
-function CreateClothingLocationRoute() {
-  return <CreateClothingLocationPage />
-}
