--- client/src/routes/clothing-management/types/$clothingTypeId/edit.tsx
@@ -1,13 +1,0 @@
-import { createFileRoute } from "@tanstack/react-router"
-
-import EditClothingTypePage from "#/clothing/components/types/edit/EditClothingTypePage"
-
-export const Route = createFileRoute(
-  "/clothing-management/types/$clothingTypeId/edit",
-)({
-  component: EditClothingTypeRoute,
-})
-
-function EditClothingTypeRoute() {
-  return <EditClothingTypePage />
-}
