--- client/src/routes/clothing-management/types/index.tsx
@@ -1,11 +1,0 @@
-import { createFileRoute } from "@tanstack/react-router"
-
-import ClothingTypesPage from "#/clothing/components/types/list/ClothingTypesPage"
-
-export const Route = createFileRoute("/clothing-management/types/")({
-  component: ClothingTypesRoute,
-})
-
-function ClothingTypesRoute() {
-  return <ClothingTypesPage />
-}
