--- client/src/routes/clothing-management/types/new.tsx
@@ -1,11 +1,0 @@
-import { createFileRoute } from "@tanstack/react-router"
-
-import CreateClothingTypePage from "#/clothing/components/types/create/CreateClothingTypePage"
-
-export const Route = createFileRoute("/clothing-management/types/new")({
-  component: CreateClothingTypeRoute,
-})
-
-function CreateClothingTypeRoute() {
-  return <CreateClothingTypePage />
-}
