--- client/src/routes/pool-clothing/checkout.tsx
@@ -1,16 +1,0 @@
-import { createFileRoute } from "@tanstack/react-router"
-
-import RoleGuard from "#/components/base/RoleGuard"
-import CheckoutPage from "#/clothing/checkout/components/CheckoutPage"
-
-export const Route = createFileRoute("/pool-clothing/checkout")({
-  component: CheckoutRoute,
-})
-
-function CheckoutRoute() {
-  return (
-    <RoleGuard allowedRoles={["USER"]}>
-      <CheckoutPage />
-    </RoleGuard>
-  )
-}
